Finalize proposal of cert download and convert
The proposed method of processing certificates is breadth-first.
By that I mean that single task is done for all certificates, then next
task is evaluated. All tasks should be implemented as static methods
and are to be paralelized by calling `sec-certs/cert_processing.py`.
Currently, a generic method `process_paralel()` sits there that is
capable of pretty much arbitrary parallel processing (both using threads
and processes). Any methods to-be called by this method must require
unpacked arguments on the input.