Allow for download through our caching proxy at sec-certs.org
Recently, CCportal changed infra and enabled a rate limit. Our
usual download procedure hits this rate limit and is able to download
only a small subset of the artifacts.

Thus a proxy is running at sec-certs.org that caches the artifacts.
This commit makes it possible to use the proxy in the tool.

By querying the proxy on the background, the server makes sure that
the rate limit is not hit and the artifacts are available.